refactor: address code review comments on metadata hyperlinks
