+++ src/app/shared/utils/make-links.ts
@@ -1,15 +1,46 @@
+import { addOperatorToFilterValue } from '../search/search.utils';
+
+/**
+ * Escape HTML special characters so that non-URL parts of a metadata value
+ * are rendered as plain text when inserted via [innerHTML].
+ */
+function escapeHtml(text: string): string {
+  return text
+    .replace(/&/g, '&amp;')
+    .replace(/</g, '&lt;')
+    .replace(/>/g, '&gt;')
+    .replace(/"/g, '&quot;')
+    .replace(/'/g, '&#039;');
+}
+
 /**
  * Convert raw text URLs into clickable HTML links.
  * Detects http, https, ftp URLs and www. addresses.
+ * Non-URL parts are HTML-escaped to prevent markup injection via [innerHTML].
  *
  * Ported from dtq-dev clarin-shared-util.ts to be reused across components.
  */
-export function makeLinks(text: string): string {
+export function makeLinks(text: string | null | undefined): string | undefined {
+  if (text == null) {
+    return undefined;
+  }
   const regex = /(?:https?|ftp):\/\/[^\s)]+|www\.[^\s)]+/g;
-  return text?.replace(regex, (url) => {
+  let result = '';
+  let lastIndex = 0;
+  let match: RegExpExecArray | null;
+
+  while ((match = regex.exec(text)) !== null) {
+    // Escape the plain text segment before this URL
+    result += escapeHtml(text.slice(lastIndex, match.index));
+    const url = match[0];
     const href = url.startsWith('www.') ? `https://${url}` : url;
-    return `<a href="${href}" target="_blank" rel="noopener noreferrer">${url}</a>`;
-  });
+    result += `<a href="${escapeHtml(href)}" target="_blank" rel="noopener noreferrer">${escapeHtml(url)}</a>`;
+    lastIndex = regex.lastIndex;
+  }
+
+  // Escape the remaining plain text after the last URL
+  result += escapeHtml(text.slice(lastIndex));
+  return result;
 }
 
 /**
@@ -32,7 +63,6 @@ const WOS_RECORD = 'https://www.webofscience.com/wos/woscc/full-record/';
 
 // --- Internal search configuration ---
 const SEARCH_PATH = '/search';
-const SEARCH_FILTER_SUFFIX = ',equals';
 
 /** Maps metadata keys to their corresponding search filter parameter. */
 const SEARCH_FIELD_FILTERS: Record<string, string> = {
@@ -52,7 +82,7 @@ const HTTP_URL_PATTERN = /^https?:\/\//i;
  * For specific metadata fields, build an appropriate hyperlink.
  * Returns null when the field needs no special treatment (fall back to makeLinks).
  */
-export function getMetadataLink(key: string, value: string): MetadataLink | null {
+export function getMetadataLink(key: string, value: string | null | undefined): MetadataLink | null {
   const trimmed = value?.trim();
   if (!trimmed) {
     return null;
@@ -74,7 +104,7 @@ export function getMetadataLink(key: string, value: string): MetadataLink | null
   // Internal search links (subject, author, …)
   const filterParam = SEARCH_FIELD_FILTERS[key];
   if (filterParam) {
-    return { external: false, routerLink: SEARCH_PATH, queryParams: { [filterParam]: `${trimmed}${SEARCH_FILTER_SUFFIX}` } };
+    return { external: false, routerLink: SEARCH_PATH, queryParams: { [filterParam]: addOperatorToFilterValue(trimmed, 'equals') } };
   }
 
   return null;
